refactor(settings): split SettingsScreen into per-section files (#42)
Extract the 12 `_*Section` widgets and their helpers from the 1,823-line
`lib/screens/settings_screen.dart` into `lib/screens/settings/sections/`.
The screen itself becomes a 50-line thin shell holding the section imports
and the `ListView.separated`.

Shared `_SectionHeader` promoted to public `SectionHeader` at
`lib/screens/settings/widgets/section_header.dart`. The full-screen
`_SmartBeaconingParamsScreen` moved to
`lib/screens/settings/smart_beaconing_params_screen.dart`
(public; state class stays private).

Single-owner helpers (`_ColorSwatch`, `_SymbolPickerTile`,
`_LocationSourcePicker`, `_IosBackgroundLocationPrompt`, `_IntervalTile`,
`_NotificationChannelTile`) remain library-private inside their owning
section file. `AboutSection.kVersion` is promoted to public so
`acknowledgements_section.dart` can surface the same version string.

Purely mechanical — no behavior change. `dart format`, `flutter analyze`,
and all 412 tests pass. Unblocks v0.13 Phase 1 (passcode row) and Phase 3
(APRS-IS Filter section) additions landing in dedicated files instead of
wedging into the monolith.

Also adds `docs/milestones/v0.13-security-connectivity-spec.md` to
`.gitignore` — the v0.13 spec is a local planning artifact per Eric's
direction.
