+++ lib/screens/settings/sections/about_section.dart
@@ -1,0 +1,107 @@
+import 'package:flutter/material.dart';
+import 'package:material_symbols_icons/symbols.dart';
+import 'package:url_launcher/url_launcher.dart';
+
+import '../../../ui/widgets/meridian_wordmark.dart';
+import '../widgets/section_header.dart';
+
+class AboutSection extends StatelessWidget {
+  const AboutSection({super.key});
+
+  static const kVersion = '0.1.0';
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        const SectionHeader('About'),
+        Padding(
+          padding: const EdgeInsets.fromLTRB(16, 4, 16, 8),
+          child: Card.filled(
+            child: Padding(
+              padding: const EdgeInsets.symmetric(vertical: 24, horizontal: 16),
+              child: Center(
+                child: Column(
+                  children: [
+                    const MeridianWordmark.horizontal(height: 64),
+                    const SizedBox(height: 16),
+                    const SizedBox(height: 4),
+                    Text(
+                      'Version $kVersion',
+                      style: theme.textTheme.bodyMedium?.copyWith(
+                        color: theme.colorScheme.onSurfaceVariant,
+                      ),
+                    ),
+                    const SizedBox(height: 2),
+                    Text(
+                      'APRS for the Modern Ham',
+                      style: theme.textTheme.bodySmall?.copyWith(
+                        color: theme.colorScheme.onSurfaceVariant,
+                      ),
+                    ),
+                    const SizedBox(height: 10),
+                    Text(
+                      '© 2026 Eric Pasch  ·  GPL v3',
+                      style: theme.textTheme.bodySmall?.copyWith(
+                        color: theme.colorScheme.outline,
+                      ),
+                    ),
+                    const SizedBox(height: 10),
+                    Text(
+                      'Inter typeface by Rasmus Andersson (SIL OFL)',
+                      style: theme.textTheme.bodySmall?.copyWith(
+                        color: theme.colorScheme.outline,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ),
+        ),
+        ListTile(
+          leading: const Icon(Icons.devices),
+          title: const Text('Device identification data'),
+          subtitle: const Text(
+            '© Heikki Hannikainen (OH7LZB) and contributors · CC BY-SA 2.0',
+          ),
+          trailing: const Icon(Symbols.open_in_new, size: 16),
+          onTap: () => _launchUri(
+            Uri.parse('https://github.com/aprsorg/aprs-deviceid'),
+            context,
+          ),
+        ),
+        ListTile(
+          leading: const Icon(Symbols.code),
+          title: const Text('GitHub'),
+          subtitle: const Text('github.com/epasch/meridian-aprs'),
+          trailing: const Icon(Symbols.open_in_new, size: 16),
+          onTap: () => _launchUri(
+            Uri.parse('https://github.com/epasch/meridian-aprs'),
+            context,
+          ),
+        ),
+        ListTile(
+          leading: const Icon(Symbols.language),
+          title: const Text('Website'),
+          subtitle: const Text('meridianaprs.com'),
+          trailing: const Icon(Symbols.open_in_new, size: 16),
+          onTap: () =>
+              _launchUri(Uri.parse('https://meridianaprs.com'), context),
+        ),
+      ],
+    );
+  }
+}
+
+Future<void> _launchUri(Uri uri, BuildContext context) async {
+  if (!await launchUrl(uri, mode: LaunchMode.externalApplication)) {
+    if (context.mounted) {
+      ScaffoldMessenger.of(
+        context,
+      ).showSnackBar(SnackBar(content: Text('Could not open ${uri.host}')));
+    }
+  }
+}
