+++ lib/screens/settings/widgets/section_header.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+/// Section header helper for the settings screen.
+class SectionHeader extends StatelessWidget {
+  const SectionHeader(this.title, {super.key});
+
+  final String title;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return Padding(
+      padding: const EdgeInsets.fromLTRB(16, 24, 16, 4),
+      child: Text(
+        title.toUpperCase(),
+        style: theme.textTheme.labelSmall?.copyWith(
+          color: theme.colorScheme.primary,
+          fontWeight: FontWeight.w700,
+          letterSpacing: 0.8,
+        ),
+      ),
+    );
+  }
+}
